added OrdersHaveProduct with extra column for units in order, using composite key,refactoring occured while testing the user commands with Postman

+++ pet_shop/src/main/java/com/example/s13firstspring/models/repositories/DeliveryRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DeliveryRepository extends JpaRepository<Delivery,Long> {
+public interface DeliveryRepository extends JpaRepository<Delivery,Integer> {
 
-    Delivery getById(long id);
+    Delivery getById(int id);
 
 }
